FROMLIST: fs: introduce SB_INLINECRYPT
Introduce SB_INLINECRYPT, which is set by filesystems that wish to use
blk-crypto for file content en/decryption. This flag maps to the
'-o inlinecrypt' mount option which multiple filesystems will implement,
and code in fs/crypto/ needs to be able to check for this mount option
in a filesystem-independent way.

Signed-off-by: Satya Tangirala <satyat@google.com>

Bug: 137270441
Link: https://lore.kernel.org/r/20200617075732.213198-2-satyat@google.com
Test: tested as series; see Ie1b77f7615d6a7a60fdc9105c7ab2200d17636a8
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: If3a5f3214e0e54b87503a448c5b0c44b2edfde0d